delete the salary parameter and change test cases for that
